Streaming interface on supernode
