+++ streaming/streamDataTypes.go
